requirements.txt
aiocontextvars==0.2.2
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
contextvars==2.4
decorator==5.0.9
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
dynaconf==3.1.4
hvac==0.11.0
idna==3.2
immutables==0.16
importlib-metadata==4.6.3
knox==0.1.11
loguru==0.5.3
Pillow==8.3.1
pytz==2021.1
requests==2.26.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
validators==0.18.2
win32-setctime==1.0.3
zipp==3.5.0
